pip-requirements.txt
setuptools
tornado
pymongo
requests
cryptography
pyOpenSSL
